# Performance improvements to init_pipeline.pl/generate_graph.pl